oo-install assistant.rb: add option to create node and district together to avoid infinite loop
Bug 1259072
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1259072
Users could get into an infinite loop when there were no nodes available to be assigned to a new district when creating a district.  Users would be forced to delete an existing district, but if that isn't the desired action, they would be put back into the choice of which node to use and district to remove.

Adds a question to ui_modify_districts which asks the user if they want to create a new node whenever they are creating a new district.  This should avoid roundabout ways of creating nodes before districts or issues where a new district is created without any nodes.  In the situation where the infinite loop can occur, the question is reworded to emphasize the consequences of not creating a node first.  Users will be able to continue without adding a node, which means there is no loss of functionality.

Users are also reminded about the 'q' option to exit the current action during the infinite loop, so users should be able to easily back out if they get into an unavoidable situation.